fix uint zero minus

+++ src/neo/Ledger/Blockchain.State.cs
@@ -66,7 +66,7 @@ namespace Neo.Ledger
 
         private void CheckStateRootCache()
         {
-            var index = Height - 1;
+            var index = (long)Height - 1;
             if (0 <= index && state_root_cache.TryGetValue(index, out StateRoot root))
             {
                 state_root_cache.Remove(index);
